requirements.txt
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
PyYAML==6.0.2
uritemplate==4.1.1
